First version DONE
Thought I would get scope creep from this project but I didn't.

This project was fun to work on daily I didn't really spend more than 1 hour a day on it; at most 3 hours; and I feel like I learned a lot especially about security, organization, and classes. The biggest reason I was not spending a lot of time on this project is because I was studying, but I'm glad I worked on it.

Now what? Well I am going to rewrite it then probably leave it I might add features in the future but its very unlikely; I don't have a issue remembering passwords.